+++ wNim/private/controls/wControl.nim
@@ -37,6 +37,8 @@
 ##   `wHyperlinkCtrl <wHyperlinkCtrl.html>`_
 ##   `wSplitter <wSplitter.html>`_
 ##   `wIpCtrl <wIpCtrl.html>`_
+##   `wWebView <wWebView.html>`_
+##   `wHotkeyCtrl <wHotkeyCtrl.html>`_
 ##   `wRebar <wRebar.html>`_  (experimental)
 
 # forward declaration